Move stubs to mock file
